requirements.txt
fastapi
uvicorn
python-jose
passlib
mysql-connector-python
